Make resource config configurable
