+++ src/main/kotlin/org/radarcns/gateway/Config.kt
@@ -1,5 +1,6 @@
 package org.radarcns.gateway;
 
+import org.radarcns.gateway.inject.ManagementPortalGatewayResources
 import java.net.URI
 
 class Config {
@@ -7,4 +8,5 @@ class Config {
     var restProxyUrl: String = "http://rest-proxy-1:8082"
     var schemaRegistryUrl: String = "http://schema-registry-1:8081"
     var managementPortalUrl: String = "http://managementportal-app:8080/managementportal/"
+    var resourceConfig: Class<*> = ManagementPortalGatewayResources::class.java
 }
